requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
httpx==0.27.0
xmltodict==0.13.0
pydantic==2.11.7
zeep[async]==4.3.1
pytest-asyncio==0.23.6
